css/css-transforms/animation/rotate-interpolation-math-functions-tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>rotate interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#propdef-rotate">
<meta name="assert" content="rotate supports animation.">
<meta name="timeout" content="long">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>
</head>

<style>
:root {
  font-size: 10px;
}

.parent {
  rotate: 90deg;
}

.target {
  width: 40px;
  height: 20px;
  background-color: grey;
  rotate: 10deg;
}

.expected {
  background-color: green;
}
</style>

<template id="target-template">
<div class="parent">
    <div class="target">Text</div>
</div>
</template>

<body>
<script>
test_interpolation({
  property: 'rotate',
  from: '100deg',
  to: 'calc(sign(20rem - 20px) * 180deg)',
}, [
  {at: -1, expect: '20deg'},
  {at: 0, expect: '100deg'},
  {at: 0.125, expect: '110deg'},
  {at: 0.875, expect: '170deg'},
  {at: 1, expect: '180deg'},
  {at: 2, expect: '260deg'}
]);

test_interpolation({
  property: 'rotate',
  from: 'calc(sign(20rem - 20px) * 100deg)',
  to: 'calc(progress(10rem from 20px to 100px) * 180deg)',
}, [
  {at: -1, expect: '20deg'},
  {at: 0, expect: '100deg'},
  {at: 0.125, expect: '110deg'},
  {at: 0.875, expect: '170deg'},
  {at: 1, expect: '180deg'},
  {at: 2, expect: '260deg'}
]);
</script>
</body>
</html>
